+++ CHANGELOG.md
@@ -8,6 +8,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `0.2.3` - August 9th, 2024
+
+### Fixed
+
+- Fixed issues with standard I/O streams on Unix systems ([#64])
+- Fixed cross-device link error when running `self-install` on Linux ([#63])
+
+[#64]: https://github.com/rojo-rbx/rokit/pull/63
+[#63]: https://github.com/rojo-rbx/rokit/pull/63
+
+## `0.2.2` - August 3rd, 2024
+
+### Added
+
+- Added a `--skip-parse` flag to `rokit authenticate` to skip token formatting checks
+
+### Fixed
+
+- Fixed `rokit authenticate` not correctly verifying certain GitHub tokens ([#60])
+
+[#60]: https://github.com/rojo-rbx/rokit/pull/60
+
 ## `0.2.1` - July 17th, 2024
 
 ### Added